fastapi-backend/requirements.txt
# Web Framework & Server
fastapi>=0.110.0
uvicorn>=0.28.0
pydantic>=2.7.0
python-dotenv>=1.0.1

# AI - Google Gemini
google-generativeai>=0.3.0

# Execution & Infrastructure
docker>=7.1.0